+++ Runtime/DiscreteDistributions/DistributionFilters/AscendantSequenceFilters/FilterProviders/AscendantSequenceFilterProvider.cs
@@ -14,13 +14,13 @@ namespace Zor.RandomGenerators.DiscreteDistributions.DistributionFilters
 
 		private AscendantSequenceFilter<T> m_sharedFilter;
 
-		public override IDisceteFilter<T> filter
+		public sealed override IDisceteFilter<T> filter
 		{
 			[Pure]
 			get => new AscendantSequenceFilter<T>(m_AscendantSequenceLength);
 		}
 
-		public override IDisceteFilter<T> sharedFilter
+		public sealed override IDisceteFilter<T> sharedFilter
 		{
 			[Pure]
 			get