+++ Runtime/DiscreteDistributions/DistributionFilters/DescendantSequenceFilters/FilterProviders/IntDescendantSequenceFilterProvider.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Zor.RandomGenerators.DiscreteDistributions.DistributionFilters
 {
 	[CreateAssetMenu(
-		menuName = CreateAssetMenuConstants.DescendantSequenceFiltersFolder +
+		menuName = CreateAssetMenuConstants.DescendantSequenceDiscreteFiltersFolder +
 			"Int Descendant Sequence Filter Provider",
 		fileName = "Int Descendant Sequence Filter Provider",
 		order = CreateAssetMenuConstants.FilterOrder