+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Extreme Sequence filter.
 - Opposite Pattern filter.
 - Pair filter.
+- Repeating Pattern filter.
 - Same Pattern filter.
 - Same Sequence filter.